moved search logic to useSearch file, implemented search on MainPage, fixed pagination if 1page in search results
